+++ pkg/innerring/processors/container/process_eacl.go
@@ -12,13 +12,13 @@ import (
 	"go.uber.org/zap"
 )
 
-func (cp *Processor) processPutEACLRequest(req container.PutContainerEACLRequest) {
+func (cp *Processor) processPutEACLRequest(req container.PutContainerEACLRequest, rfc6979 bool) {
 	if !cp.alphabetState.IsAlphabet() {
 		cp.log.Info("non alphabet mode, ignore set EACL")
 		return
 	}
 
-	err := cp.checkSetEACL(req)
+	err := cp.checkSetEACL(req, rfc6979)
 	if err != nil {
 		cp.log.Error("set EACL check failed",
 			zap.Error(err),
@@ -30,7 +30,7 @@ func (cp *Processor) processPutEACLRequest(req container.PutContainerEACLRequest
 	cp.approveSetEACL(req)
 }
 
-func (cp *Processor) checkSetEACL(req container.PutContainerEACLRequest) error {
+func (cp *Processor) checkSetEACL(req container.PutContainerEACLRequest, rfc6979 bool) error {
 	// unmarshal table
 	table, err := eacl.Unmarshal(req.EACL)
 	if err != nil {
@@ -64,8 +64,9 @@ func (cp *Processor) checkSetEACL(req container.PutContainerEACLRequest) error {
 		idContainerSet:  true,
 		idContainer:     idCnr,
 		binTokenSession: req.SessionToken,
-		binPublicKey:    req.VerificationScript,
-		signature:       req.InvocationScript,
+		verifScript:     req.VerificationScript,
+		invocScript:     req.InvocationScript,
+		rfc6979:         rfc6979,
 		signedData:      req.EACL,
 	})
 	if err != nil {