+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changelog for NeoFS Node
 - IR supports recently added Container contract's `create`, `remove` and `putEACL` methods now (#3282)
 - SN attempts to make notary requests calling Container contract's `create`, `remove` and `putEACL` methods (#3282)
 - SN listens and handles recently added Container contract's `Created` and `Removed` notifications (#3282)
+- IR now supports N3 account authorization of Container contract's `create`, `remove` and `putEACL` ops (#3298)
 
 ### Fixed
 - Bearer token signed not by its issuer is no longer passed (#3216)